requirements.txt
streamlit>=1.22.0
pdfplumber>=0.7.6
nltk>=3.8.1
sentence-transformers>=2.2.2
numpy>=1.24.3
faiss-cpu>=1.7.4
transformers>=4.28.1
openai>=1.2.0
tenacity>=8.2.2
python-dotenv>=1.0.0
